% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dygraph.R
\name{dygraph}
\alias{dygraph}
\alias{dygraph.default}
\alias{dygraph.data.frame}
\alias{dygraph.grouped_df}
\alias{dygraph.tbl_ts}
\title{Methods for creating dygraphs from tidy data}
\usage{
dygraph(data, ..., time = NULL)

\method{dygraph}{default}(data, ...)

\method{dygraph}{data.frame}(data, ..., time = NULL)

\method{dygraph}{grouped_df}(data, ..., time = NULL)

\method{dygraph}{tbl_ts}(data, ...)
}
\arguments{
\item{data}{The data}

\item{...}{Unquoted column name(s) of the desired series and any named
arguments to \code{\link[dygraphs]{dygraph}}}

\item{time}{The unquoted column name containing the time index. If not
supplied the first column identified by
\code{\link[lubridate]{is.timepoint}} will be used}
}
\value{
An interactive dygraph plot that can be modified with any other the
functions in the package dygraphs.
}
\description{
For many tidy datasets it is not convenient to use
\code{\link[dygraphs]{dygraph}} to create quick, interactive time series
plots. Methods are supplied for dataframes, tibbles, grouped tibbles, and
tsibbles. Currently the time column must be specified if data is of any of
the classes above.
}
\details{
For any data object that is not a dataframe, tibble, grouped tibble, or
tsibble \code{\link[dygraphs]{dygraph}} is called directly. Otherwise a
dygraph is generated with a separate series for each column passed to
\code{...}. Additionally separate series are made for each unique group for
grouped tibbles and tsibbles (which are grouped based on key values).

Be aware that because a separate series is made for each group and each
measure the plot can become overly busy quickly if too many groups or
measures are passed to the function.
}
